backend/requirements.txt
fastapi>=0.115,<0.120
starlette>=0.40,<0.42
uvicorn==0.29.0
pydantic==2.6.4
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-http>=1.24.0
chromadb==0.4.24
numpy<2.0.0
yfinance
requests
openai
supabase
google-genai>=1.30,<2.0
google-auth
PyJWT
pandas
apscheduler
aiofiles
huggingface_hub
pyarrow
httpx>=0.28,<0.29
python-dotenv>=1.0.0
PyYAML>=6.0
